change restructure, split interface
